Refactor: extract RankService
